+++ NEWS.md
@@ -5,6 +5,7 @@
 * Updates to account for changes in {forcats} v1.0.0 and {dplyr} v1.1.0.
 
 * Updated `style_ratio()` to now format negative values.
+* `add_ci.tbl_svysummary()` now takes properly into account the `percent` argument (#1470)
 
 # gtsummary 1.7.0